fix draw label

+++ Common/Panel.h
@@ -8,7 +8,7 @@ public:
 	Panel();
 	~Panel();
 
-	void draw(Graphics& g, int x, int y, size_t z);
+	void draw(Graphics& g, short x, short y, size_t z);
 
 	void Add(Control* control);
 	//void getAllControls(vector<Control*>* Controls) { controls= Controls; }
